requirements.txt
django
Pillow
djangorestframework
psycopg2-binary
django-allauth